// src/app/shared/dtos/request/employee/update-employee-dto.ts
export class UpdateEmployeeDTO {
  employeeId: any;
  name: string = "";
  photoUrl: string = "";
  phone: string = "";
  role: string = "";
  employeeAddressId: any;
  address: string = "";
  number: string = "";
  complement: string = "";
  neighborhood: string = "";
  cep: string = "";
  city: string = "";
  state: string = "";
}